# swagger.yaml
openapi: 3.0.0
info:
  title: Video Generator Microservice API
  description: API para generación de contenido multimedia (audio, imágenes y videos)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
paths:
...
  /media/image:
    post:
      summary: Generar imagen promocional
      description: Genera imágenes promocionales (puede usar DALL·E o FLUX)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  example: "Un zapato deportivo flotando sobre un fondo futurista azul"
                plan:
                  type: string
                  enum: [FREE, CREATOR, PRO]
                  example: "CREATOR"
                useFlux:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Imagen generada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: ✅ Imagen generada correctamente
                  result:
                    type: object
                    properties:
                      imageUrl:
                        type: string
                        example: https://storageaccount.blob.core.windows.net/images/image.png?sv=2020-08-04&ss=bfqt&srt=sco&sp=rwdlacupitfx&se=2025-10-16T00%3A00%3A00Z&st=2025-10-15T00%3A00%3A00Z&spr=https&sig=token
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor
  /media/flux-image:
    post:
      summary: Generar imagen con FLUX-1.1-pro
      description: Genera imágenes usando FLUX-1.1-pro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  example: "A majestic lion in the savannah at sunset"
                plan:
                  type: string
                  enum: [FREE, CREATOR, PRO]
                  example: "CREATOR"
                size:
                  type: string
                  enum: [1024x1024, 1024x768, 768x1024]
                  example: "1024x1024"
                isJsonPrompt:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Imagen generada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: ✅ FLUX image generated successfully
                  data:
                    type: object
                    properties:
                      imageUrl:
                        type: string
                        example: https://storageaccount.blob.core.windows.net/images/image.png?sv=2020-08-04&ss=bfqt&srt=sco&sp=rwdlacupitfx&se=2025-10-16T00%3A00%3A00Z&st=2025-10-15T00%3A00%3A00Z&spr=https&sig=token
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor
...